Prototyping/SEC-Edgar/pyproject.toml
[project]
name = "sec-edgar-filings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "camelot-py[cv]>=1.0.9",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pdfplumber>=0.11.9",
    "tabula-py[jpype]>=2.10.0",
]
